Move job utils to legacy package
Currently the V1 CLI commands that we want to move to V2 (docker run,
wasm run) make use of the pkg/job package to build a job ready for
submission.  The files here make use of pkg/model.

To make it clearer what needs to change in the CLI this PR moves pkg/job
to pkg/legacyjob. This leaves pkg/job open to re-implementation
following a similar but not identical approach (builders with options).

This aims to make it easier to move over the job creation commands
_without_ forcing the implementor to change both, and all the tests in
the same PR. Now it should be possible for docker run to move over to v2
independently of the wasm run command.
