+++ cmd/util/execute.go
@@ -6,7 +6,7 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/bacalhau-project/bacalhau/cmd/util/flags/cliflags"
-	"github.com/bacalhau-project/bacalhau/pkg/job"
+	legacy_job "github.com/bacalhau-project/bacalhau/pkg/legacyjob"
 	"github.com/bacalhau-project/bacalhau/pkg/model"
 )
 
@@ -15,7 +15,7 @@ func ExecuteJob(ctx context.Context,
 	j *model.Job,
 	runtimeSettings *cliflags.RunTimeSettingsWithDownload,
 ) (*model.Job, error) {
-	err := job.VerifyJob(ctx, j)
+	err := legacy_job.VerifyJob(ctx, j)
 	if err != nil {
 		log.Ctx(ctx).Err(err).Msg("Job failed to validate.")
 		return nil, err
