+++ pkg/legacyjob/script_checker.go
@@ -1,4 +1,4 @@
-package job
+package legacyjob
 
 import (
 	"fmt"
