+++ pkg/test/scenario/job_checkers.go
@@ -1,7 +1,7 @@
 package scenario
 
 import (
-	"github.com/bacalhau-project/bacalhau/pkg/job"
+	legacy_job "github.com/bacalhau-project/bacalhau/pkg/legacyjob"
 
 	"github.com/bacalhau-project/bacalhau/pkg/model"
 )
@@ -9,12 +9,12 @@ import (
 // WaitUntilSuccessful returns a set of job.CheckStatesFunctions that will wait
 // until the job they are checking reaches the Completed state on the passed
 // number of nodes. The checks will fail if any job errors.
-func WaitUntilSuccessful(nodes int) []job.CheckStatesFunction {
-	return []job.CheckStatesFunction{
-		job.WaitExecutionsThrowErrors([]model.ExecutionStateType{
+func WaitUntilSuccessful(nodes int) []legacy_job.CheckStatesFunction {
+	return []legacy_job.CheckStatesFunction{
+		legacy_job.WaitExecutionsThrowErrors([]model.ExecutionStateType{
 			model.ExecutionStateFailed,
 		}),
-		job.WaitForExecutionStates(map[model.ExecutionStateType]int{
+		legacy_job.WaitForExecutionStates(map[model.ExecutionStateType]int{
 			model.ExecutionStateCompleted: nodes,
 		}),
 	}
